+++ src/pictex/nodes/node.py
@@ -4,22 +4,17 @@ from typing import Optional, Tuple
 import skia
 from ..models import Style, Shadow, PositionMode, RenderProps, CropMode
 from ..painters import Painter
-from ..utils import create_composite_shadow_filter, clone_skia_rect, to_int_skia_rect
+from ..utils import create_composite_shadow_filter, clone_skia_rect, to_int_skia_rect, cached_property, Cacheable
 from ..layout import SizeResolver
 
-class Node:
+class Node(Cacheable):
 
     def __init__(self, style: Style):
+        super().__init__()
         self._raw_style = style
         self._parent: Optional[Node] = None
         self._children: list[Node] = []
-        self._computed_styles: Optional[Style] = None
-        self._size: Optional[Tuple[int, int]] = None
-        self._content_bounds: Optional[skia.Rect] = None
-        self._padding_bounds: Optional[skia.Rect] = None
-        self._border_bounds: Optional[skia.Rect] = None
-        self._margin_bounds: Optional[skia.Rect] = None
-        self._paint_bounds: Optional[skia.Rect] = None
+        self._forced_size: Tuple[Optional[int], Optional[int]] = (None, None)
         self._render_props: Optional[RenderProps] = None
         self._absolute_position: Optional[Tuple[float, float]] = None
 
@@ -31,17 +26,21 @@ class Node:
     def children(self) -> list[Node]:
         return self._children
 
-    @property
+    @cached_property()
     def computed_styles(self) -> Style:
-        if self._computed_styles is None:
-            self._computed_styles = self._compute_styles()
-        return self._computed_styles
+        return self._compute_styles()
 
-    @property
+    @cached_property(group='bounds')
     def size(self) -> Tuple[int, int]:
-        if self._size is None:
-            self._size = (self.border_bounds.width(), self.border_bounds.height())
-        return self._size
+        return (self.border_bounds.width(), self.border_bounds.height())
+    
+    @cached_property(group='bounds')
+    def content_width(self) -> int:
+        return SizeResolver(self).resolve_width()
+
+    @cached_property(group='bounds')
+    def content_height(self) -> int:
+        return SizeResolver(self).resolve_height()
 
     @property
     def absolute_position(self) -> Optional[Tuple[float, float]]:
@@ -68,35 +67,25 @@ class Node:
         self._absolute_position = position.get_relative_position(self_width, self_height, root_width, root_height)
         return self._absolute_position
 
-    @property
+    @cached_property(group='bounds')
     def padding_bounds(self):
-        if self._padding_bounds is None:
-            self._padding_bounds = to_int_skia_rect(self._compute_padding_bounds())
-        return self._padding_bounds
+        return to_int_skia_rect(self._compute_padding_bounds())
 
-    @property
+    @cached_property(group='bounds')
     def border_bounds(self):
-        if self._border_bounds is None:
-            self._border_bounds = to_int_skia_rect(self._compute_border_bounds())
-        return self._border_bounds
+        return to_int_skia_rect(self._compute_border_bounds())
 
-    @property
+    @cached_property(group='bounds')
     def margin_bounds(self):
-        if self._margin_bounds is None:
-            self._margin_bounds = to_int_skia_rect(self._compute_margin_bounds())
-        return self._margin_bounds
+        return to_int_skia_rect(self._compute_margin_bounds())
 
-    @property
+    @cached_property(group='bounds')
     def content_bounds(self) -> skia.Rect:
-        if self._content_bounds is None:
-            self._content_bounds = to_int_skia_rect(self._compute_content_bounds())
-        return self._content_bounds
+        return to_int_skia_rect(skia.Rect.MakeWH(self.content_width, self.content_height))
 
-    @property
+    @cached_property(group='bounds')
     def paint_bounds(self) -> skia.Rect:
-        if self._paint_bounds is None:
-            self._paint_bounds = to_int_skia_rect(self._compute_paint_bounds())
-        return self._paint_bounds
+        return to_int_skia_rect(self._compute_paint_bounds())
 
     def _compute_padding_bounds(self) -> skia.Rect:
         """
@@ -140,20 +129,6 @@ class Node:
             border_bounds.bottom() + margin.bottom
         )
 
-    def _compute_content_bounds(self) -> skia.Rect:
-        """
-        Compute the inner content bounds (implicit), relative to the node box size, (0, 0).
-        Implicit means that it ignores the explicit size set from the styles for the node.
-        """
-        return SizeResolver(self).resolve()
-
-    def _compute_intrinsic_content_bounds(self) -> skia.Rect:
-        """
-        Compute the intrinsic content bounds. That is, ignoring any size strategy set.
-        It measures the actual content (if the strategy is 'fit-content', then it's the same that _compute_content_bounds())
-        """
-        raise NotImplementedError("_compute_implicit_content_bounds() is not implemented")
-
     def _compute_paint_bounds(self) -> skia.Rect:
         """
         Compute the paint bounds, including anything that will be painted for this node, even outside the box (like shadows).
@@ -163,6 +138,20 @@ class Node:
 
     def _get_painters(self) -> list[Painter]:
         raise NotImplementedError("_get_painters() is not implemented")
+    
+    def compute_intrinsic_width(self) -> skia.Rect:
+        """
+        Compute the intrinsic width. That is, ignoring any size strategy set.
+        It measures the actual content (if the strategy is 'fit-content', then it's the same that self.content_width)
+        """
+        raise NotImplementedError("compute_intrinsic_width() is not implemented")
+    
+    def compute_intrinsic_height(self) -> skia.Rect:
+        """
+        Compute the intrinsic height. That is, ignoring any size strategy set.
+        It measures the actual content (if the strategy is 'fit-content', then it's the same that self.content_height)
+        """
+        raise NotImplementedError("compute_intrinsic_height() is not implemented")
 
     def prepare_tree_for_rendering(self, render_props: RenderProps) -> None:
         """
@@ -219,15 +208,25 @@ class Node:
         for child in self._children:
             child.clear()
 
-        self._computed_styles = None
-        self._size = None
-        self._content_bounds = None
-        self._padding_bounds = None
-        self._border_bounds = None
-        self._margin_bounds = None
-        self._paint_bounds = None
         self._render_props = None
         self._absolute_position = None
+        self._forced_size = (None, None)
+        self.clear_cache()
+
+    def clear_bounds(self):
+        """
+        Resets only the calculated layout and bounds information.
+        This is a more targeted version of clear().
+        """
+        for child in self._children:
+            child.clear()
+
+        self._forced_size = (None, None)
+        self.clear_cache('bounds')
+
+    def _set_forced_size(self, width: Optional[int] = None, height: Optional[int] = None) -> None:
+        """Allows a parent to impose a size on this node."""
+        self._forced_size = (width, height)
 
     def _compute_styles(self) -> Style:
         parent_computed_styles = self._parent.computed_styles if self._parent else None